+++ src/main/java/com/softlab/okr/entity/KeyUser.java
@@ -13,7 +13,7 @@ import java.util.Date;
 
 /**
  * <p>
- * 
+ *
  * </p>
  *
  * @author Mybatis-plus自动生成
@@ -31,7 +31,7 @@ public class KeyUser {
      * 主键
      */
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
     /**
      * 钥匙id
      */
